clean up null handling warnings

+++ ClinicalTrials.Apps/Queries.xaml.cs
@@ -35,7 +35,7 @@ namespace ClinicalTrials.Apps
 
         private async void RefreshQueries_Clicked(object sender, EventArgs e)
         {
-            fetchButton.Text = "⏱️...";
+            fetchButton.Text = "⏱️Fetch All";
             foreach (var queryFileObj in queryView.ItemsSource)
             {
                 var queryFile = queryFileObj as QueryFile;
@@ -74,8 +74,11 @@ namespace ClinicalTrials.Apps
         private void DeleteItem_Invoked(object sender, EventArgs e)
         {
             var item = (sender as BindableObject)?.BindingContext as QueryFile;
-            queryFiles.Remove(item);
-            DeviceProfileUtility.DeleteItem(item.Name);
+            if (item != null)
+            {
+                queryFiles.Remove(item);
+                DeviceProfileUtility.DeleteItem(item.Name);
+            }
         }
     }
 }
